pyproject.toml
[tool.poetry]
authors = ["long2ice <long2ice@gmail.com>"]
description = "A fast admin dashboard based on FastAPI and TortoiseORM with tabler ui, inspired by Django admin."
documentation = "https://fastapi-admin.github.io"
homepage = "https://github.com/fastapi-admin/fastapi-admin"
keywords = ["fastapi", "admin", "dashboard", "tortoise-orm"]
license = "Apache-2.0"
name = "fastapi-admin"
packages = [
    { include = "fastapi_admin" },
]
readme = "README.md"
repository = "https://github.com/fastapi-admin/fastapi-admin.git"
version = "1.0.5"

[tool.poetry.dependencies]
Babel = "*"
aiofiles = "*"
redis = "^4.2.0rc1"
bcrypt = "*"
fastapi = "*"
jinja2 = "*"
python = "^3.8"
python-multipart = "*"
tortoise-orm = "*"
uvicorn = { version = "*", extras = ["standard"] }
pendulum = "*"
fastapi-login = "1.9.3"

[tool.poetry.dev-dependencies]
# test
pytest = "*"
pytest-asyncio = "*"
pytest-mock = "*"
pytest-xdist = "*"
# lint
black = "*"
flake8 = "*"
isort = "*"
pylint = "*"
# example
asyncmy = "*"
asyncpg = "*"
python-dotenv = "*"
mypy = "*"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "setuptools"]
